microservices/kubeApi/pyproject.toml
[tool.poetry]
name = "gwa-kube-api"
version = "1.0.0"
description = "An API to manage Kubernetes and OCP resources"
authors = ["Nithin Shekar Kuruba <nithinshekar.kuruba@gov.bc.ca>"]

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "^0.68.1"
uvicorn = "^0.15.0"
PyYAML = "^5.4.1"
python-keycloak = "^0.26.1"
Authlib = "^0.15.4"
httpx = "^0.19.0"
fastapi-utils = "^0.2.1"

[tool.poetry.dev-dependencies]
python-dotenv = "^0.19.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
